% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{st_save_final}
\alias{st_save_final}
\title{Save final model to an RDS file}
\usage{
st_save_final(final_model, file)
}
\arguments{
\item{final_model}{object returned by \code{\link[=st_fit_final]{st_fit_final()}}}

\item{file}{path to \code{.rds} file}
}
\value{
invisibly returns \code{file}
}
\description{
Save final model to an RDS file
}
\examples{
# st_save_final(final, "spectrotune_final_model.rds")
}
